requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
django-qr-code==4.0.1
djangorestframework==3.15.1
importlib-metadata==7.1.0
mysqlclient==2.2.4
pillow==10.3.0
pydantic==2.7.1
pydantic-core==2.18.2
pypng==0.20220715.0
qrcode==7.4.2
segno==1.6.1
sqlparse==0.5.0
typing-extensions==4.12.0
zipp==3.18.2
